+++ CHANGELOG.md
@@ -9,10 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+### **Added**
+
+- added `sagemaker-custom-kernel` module
+- added batch inference project template to `sagemaker-templates-service-catalog` module
 - added EFS removal policy to `mlflow-fargate` module
-- added `batch_inference` template to the `sagemaker-templates-service-catalog` module
-- remove AmazonSageMakerFullAccess from `multi_account_basic` template in the `sagemaker-templates-service-catalog` module
+
+### **Changed**
+
 - update MySQL instance to use T3 instance type
+- upgrade `cdk_ecr_deployment` version to fix the deprecated `go1.x` lambda runtime
+
+### **Removed**
+
+- remove AmazonSageMakerFullAccess from `multi_account_basic` template in the `sagemaker-templates-service-catalog` module
 - remove AmazonSageMakerFullAccess from `sagemaker-endpoint` module
 
 ## v1.0.0